requirements.txt
python-socketio[client]==5.11.2
websocket-client==1.8.0
requests==2.32.3
python-dotenv==1.0.1
